Classify land cover based on multi-source imagery datasets

Generate semantic land cover segmentation based on map tiles and bands of Sentinel-2-l2a images by sampling data over the landcover of ESA WorldCover dataset and using random forest. The semantic segmentation includes the same classes as ESA WorldCover dataset with additional class of building (See details).

Usage

lc_sem_seg(
  bbox = NULL,
  place = NULL,
  label_year = NULL,
  tiles = c("esri", "eox"),
  tile_year = NULL,
  zoom = 17,
  s2a = NULL,
  datetime = c(),
  cloud_cover = 10,
  vege_perc = 0,
  sample_size = 10000,
  trian_split = 0.7,
  test = TRUE,
  cores = TRUE
)

Arguments

bbox

sf, sfc, or a numeric vector (xmin, ymin, xmax, ymax) defining the area of interest. Optional if place is provided.

place

character or vector. (optional) A single line address, e.g. ("1600 Pennsylvania Ave NW, Washington") or a vector of addresses (c("Madrid", "Barcelona")).

label_year

numeric. The year of interest: 2020 or 2021 for retrieving land cover from ESA WorldCover dataset. The default is 2021.

tiles

character or vector of character. One of "eox" and c("eox","esri") (default).

tile_year

numeric. The year (available from 2018 to 2024) of map tiles provided by EOX.

zoom

numeric. Zoom level of map tiles. The default is 17.

s2a

character. One of "all_bands", "ndvi", and NULL. If s2a = "all_bands", all bands of Sentinel-2-l2a will downloaded. If s2a = "ndvi", only band 4 and band 8 will be downloaded to calculate NDVI to train randomforest model. If s2a = NULL (default), Sentinel-2-l2a will not be included.

datetime

numeric vector of 2. The time of interest such as c("2020-08-01", "2020-09-01") for retrieving Sentinel-2-l2a images. (Only required, when bands = "all_bands" or bands = "ndvi")

cloud_cover

numeric. The percentage of cloud coverage for retrieving Sentinel-2-l2a images. (Only required, when bands = "all_bands" or bands = "ndvi")

vege_perc

numeric. The percentage of cloud coverage for retrieving Sentinel-2-l2a images. (Only required, when bands = "all_bands" or bands = "ndvi")

sample_size

numeric. The total number of locations to be sampled over land cover classes. The default is 10000.

trian_split

numeric. The percentage of samples used for training model. The default is 0.7.

test

logical. Whether to test the performance of randomforests model. The default is TRUE.

cores

logical. Whether parallel processing will be used to generate land cover.

Value

If test = FALSE, a land cover raster. If test = TRUE, a list includes a land cover raster, model performance table, and predictor importance plot.

Details

This function uses ESA WorldCover classification with additional class of building :

Code	Land Cover Class	RGB Color	Description

10	Tree cover	(0, 100, 0)	Geographic areas dominated by trees with a cover of 10% or more. Includes plantations, mangroves.

20	Shrubland	(255, 187, 34)	Areas dominated by shrubs with cover ≥10%. Includes evergreen/deciduous.

30	Grassland	(255, 255, 76)	Dominated by herbaceous plants with cover ≥10%. Includes savannas, pastures, steppes.

40	Cropland	(240, 150, 255)	Areas with annual crops sown/planted and harvested within 12 months.

50	Built-up	(250, 0, 0)	Built environment: buildings, roads, other man-made structures.

60	Bare/sparse vegetation	(180, 180, 180)	Lands with exposed soil/sand/rock and <10% vegetation.

70	Snow and Ice	(240, 240, 240)	Areas permanently covered by snow or glaciers.

80	Permanent water bodies	(0, 100, 200)	Lakes, rivers, reservoirs covered with water >9 months/year.

90	Herbaceous wetland	(0, 150, 160)	Wetlands with permanent/seasonal flooding and herbaceous vegetation.

95	Mangroves	(0, 207, 117)	Salt-tolerant trees/shrubs in coastal intertidal zones.

100	Moss and lichen	(250, 230, 160)	Lichen/moss-dominated lands, often with symbiotic fungi and algae.

105	Buildings	(0, 0, 0)	Buildings

Examples

sem <- lc_sem_seg(
   # bbox = c(-83.087174,42.333373,-83.042542,42.358748),
   tiles = c('esri', 'eox'),
   label_year = 2021,
   tile_year = 2024
)